backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
pypdf2==3.0.1
python-dotenv==1.0.0
scikit-learn==1.3.2
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
